feat(portal): inline MessagesSection shared section (overview returns messageToken)
